Update delve-into-go.slide
Below are some images / GIF that you could consider.  I've realized that I'm not very good at picking those out.  But I'll keep looking, and happy to collaborate or offer feedback on photos that you choose.

I changed the order as we discussed.  And I consolidated some of the slides.  I liked the timeline idea - but I'm not sure how that would look on one slide.  I imagine you'd have to do it in photoshop or something first.  Maybe ask one of the designers about it.

As I mentioned a lot of the slides have words that you can just speak, and aren't necessary to be stand alone slides.

Let me know if you want to practice on me early next week, or get any other feedback.


* I'll get to the technical details...
http://blogdailyherald.com/wp-content/uploads/2013/03/Fireside-Flog.jpg
http://todaymade.com/blog/wp-content/uploads/2012/08/typewriter-story.jpg

* It all began...
http://rest-ministries.org/wp-content/uploads/2013/02/winter-snow-tree-blizzard.jpg
http://funny-pics-fun.com/wp-content/uploads/Funny-Animals-In-The-Snow-15-320x240.jpg
<iframe src="//giphy.com/embed/QG1XKRLHhLgu4" width="480" height="208" frameBorder="0" style="max-width: 100%" class="giphy-embed" webkitAllowFullScreen mozallowfullscreen allowFullScreen></iframe>